Refactor: DND 드래그 컴포넌트 조금 더 세분화
